app/requirements.txt
folium
fastapi
uvicorn
pillow
python-multipart
aiofiles
